separate Field models to a separate module
